Split out GHasKey, add Monoid instance for List
