+++ generic-lens.cabal
@@ -53,6 +53,7 @@ library
                     , Data.Generics.Product.Internal.Positions
                     , Data.Generics.Product.Internal.Subtype
                     , Data.Generics.Product.Internal.List
+                    , Data.Generics.Product.Internal.Keyed
 
   build-depends:      base        >= 4.9 && <= 5.0
                     , profunctors >= 5.0 && <= 6.0
